Some issues were fixed
