+++ .gitignore
@@ -7,6 +7,11 @@
 __pycache__/
 *.py[cod]
 *$py.class
+dist/
+build/
+
+# Egg files
+*.egg*
 
 # PEP 582; used by e.g. github.com/David-OConnor/pyflow and github.com/pdm-project/pdm
 __pypackages__/
